Extracts useMeQuery to own hook
