+++ apps/web/components/booking/BookingListItem.tsx
@@ -12,9 +12,9 @@ import { Dialog, DialogClose, DialogContent, DialogFooter, DialogHeader } from "
 import { TextArea } from "@calcom/ui/form/fields";
 
 import { HttpError } from "@lib/core/http/error";
+import useMeQuery from "@lib/hooks/useMeQuery";
 import { inferQueryOutput, trpc } from "@lib/trpc";
 
-import { useMeQuery } from "@components/Shell";
 import { RescheduleDialog } from "@components/dialog/RescheduleDialog";
 import TableActions, { ActionType } from "@components/ui/TableActions";
