--- apps/web/components/team/currentUser.tsx
@@ -1,7 +1,0 @@
-import { useMeQuery } from "@components/Shell";
-
-export const useCurrentUser = () => {
-  const query = useMeQuery();
-  const user = query.data;
-  return user?.id;
-};
